fix : 소셜 로그인 추상화 작업
